Introduce path specifier in routes for parameter encode-decoding
It is a simple and comfortable solution, but it does not cover all the
possibilities. Slashes can still cause problems and unexpected behaviours,
but at least no input breaks the request. The next step is going to be manually
written encoder-decoder functions - which is a great effort for a little fix.